+++ .space.kts
@@ -24,7 +24,10 @@ job("continuous") {
 
     container("mcr.microsoft.com/dotnet/sdk:6.0") {
         shellScript {
-            content = "./build.sh Test"
+            content = """
+                git submodule update --init --recursive
+                ./build.sh Test
+            """
         }
     }