+++ source/Nuke.Common.Tests/CI/ConfigurationGenerationTest.cs
@@ -110,6 +110,7 @@ namespace Nuke.Common.Tests.CI
                         BranchesExcept = new[] { "except_branch" },
                         SkipTags = true,
                         SkipBranchesWithPullRequest = true,
+                        Submodules = true,
                         Secrets = new[] { "GitHubToken" }
                     }
                 );
@@ -148,7 +149,9 @@ namespace Nuke.Common.Tests.CI
                         OnPullRequestIncludePaths = new[] { "pull_request_include_path" },
                         OnPullRequestExcludePaths = new[] { "pull_request_exclude_path/**" },
                         OnWorkflowDispatchOptionalInputs = new[] { "OptionalInput" },
-                        OnWorkflowDispatchRequiredInputs = new[] { "RequiredInput" }
+                        OnWorkflowDispatchRequiredInputs = new[] { "RequiredInput" },
+                        Submodules = GitHubActionsSubmodules.Recursive,
+                        FetchDepth = 2
                     }
                 );