+++ source/Nuke.Common/Assert.cs
@@ -19,7 +19,7 @@ namespace Nuke.Common
         [ContractAnnotation("=> halt")]
         public static void Fail(string message, Exception exception = null)
         {
-            throw new ApplicationException(message, exception);
+            throw new Exception(message, exception);
         }
 
         [AssertionMethod]
@@ -80,7 +80,7 @@ namespace Nuke.Common
             {
                 throw new ArgumentException(
                     message ?? $"Expected object of type '{typeof(T).FullName}' to be not null",
-                    argumentExpression);
+                    message == null ? argumentExpression : null);
             }
 
             return obj;
@@ -115,7 +115,7 @@ namespace Nuke.Common
         [AssertionMethod]
         public static void NotEmpty<T>(
             [AssertionCondition(AssertionConditionType.IS_NOT_NULL)] [CanBeNull]
-            ICollection<T> collection,
+            IReadOnlyCollection<T> collection,
             string message = null,
             [CallerArgumentExpression("collection")]
             string argumentExpression = null)
@@ -127,7 +127,7 @@ namespace Nuke.Common
         [AssertionMethod]
         public static void Empty<T>(
             [AssertionCondition(AssertionConditionType.IS_NOT_NULL)] [CanBeNull]
-            ICollection<T> collection,
+            IReadOnlyCollection<T> collection,
             string message = null,
             [CallerArgumentExpression("collection")]
             string argumentExpression = null)
@@ -139,20 +139,20 @@ namespace Nuke.Common
         [AssertionMethod]
         public static void Count<T>(
             [AssertionCondition(AssertionConditionType.IS_NOT_NULL)] [CanBeNull]
-            ICollection<T> collection,
+            IReadOnlyCollection<T> collection,
             int length,
             string message = null,
             [CallerArgumentExpression("collection")]
             string argumentExpression = null)
         {
             if (collection.NotNull(argumentExpression: argumentExpression).Count != length)
-                throw new ArgumentException(message ?? $"Expected collection to have length of {length}");
+                throw new ArgumentException(message ?? $"Expected collection to have length of {length}", message == null ? argumentExpression : null);
         }
 
         [AssertionMethod]
         public static void HasSingleItem<T>(
             [AssertionCondition(AssertionConditionType.IS_NOT_NULL)] [CanBeNull]
-            ICollection<T> collection,
+            IReadOnlyCollection<T> collection,
             string message = null,
             [CallerArgumentExpression("collection")]
             string argumentExpression = null)
@@ -163,13 +163,13 @@ namespace Nuke.Common
         public static void FileExists(string path, string message = null, [CallerArgumentExpression("path")] string argumentExpression = null)
         {
             if (!File.Exists(path.NotNull(argumentExpression)))
-                throw new ArgumentException(message ?? $"Expected file to exist: {path}");
+                throw new ArgumentException(message ?? $"Expected file to exist: {path}", message == null ? argumentExpression : null);
         }
 
         public static void DirectoryExists(string path, string message = null, [CallerArgumentExpression("path")] string argumentExpression = null)
         {
             if (!Directory.Exists(path.NotNull(argumentExpression)))
-                throw new ArgumentException(message ?? $"Expected directory to exist: {path}");
+                throw new ArgumentException(message ?? $"Expected directory to exist: {path}", message == null ? argumentExpression : null);
         }
     }
 }