+++ source/Nuke.Common/Execution/Host.Theming.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Linq;
+using Nuke.Common.Execution;
 using Nuke.Common.Execution.Theming;
 using Nuke.Common.Utilities;
 
@@ -13,41 +14,36 @@ namespace Nuke.Common
     {
         private const string SevereMessagesOutputTemplate = "[{Level:u3}] {ExecutingTarget}: {Message:l}{NewLine}";
 
-        internal static IHostTheme DefaultTheme { get; } =
-            Environment.GetEnvironmentVariable("TERM") is { } term && term.StartsWithOrdinalIgnoreCase("xterm")
-                ? AnsiConsoleHostTheme.Default256AnsiColorTheme
-                : SystemConsoleHostTheme.DefaultSystemColorTheme;
-
         internal static string DefaultOutputTemplate => "{Timestamp:HH:mm:ss} [{Level:u3}] {Message:l}{NewLine}{Exception}";
 
         internal static void Success(string text = null)
         {
-            (Instance?.Theme ?? DefaultTheme).WriteSuccess(text);
+            (Instance?.Theme ?? Logging.DefaultTheme).WriteSuccess(text);
         }
 
         internal static void Verbose(string text = null)
         {
-            (Instance?.Theme ?? DefaultTheme).WriteVerbose(text);
+            (Instance?.Theme ?? Logging.DefaultTheme).WriteVerbose(text);
         }
 
         internal static void Debug(string text = null)
         {
-            (Instance?.Theme ?? DefaultTheme).WriteDebug(text);
+            (Instance?.Theme ?? Logging.DefaultTheme).WriteDebug(text);
         }
 
         internal static void Information(string text = null)
         {
-            (Instance?.Theme ?? DefaultTheme).WriteInformation(text);
+            (Instance?.Theme ?? Logging.DefaultTheme).WriteInformation(text);
         }
 
         internal static void Warning(string text = null)
         {
-            (Instance?.Theme ?? DefaultTheme).WriteWarning(text);
+            (Instance?.Theme ?? Logging.DefaultTheme).WriteWarning(text);
         }
 
         internal static void Error(string text = null)
         {
-            (Instance?.Theme ?? DefaultTheme).WriteError(text);
+            (Instance?.Theme ?? Logging.DefaultTheme).WriteError(text);
         }
     }
 }