+++ source/Nuke.GlobalTool/Program.Cake.cs
@@ -29,6 +29,8 @@ namespace Nuke.GlobalTool
         [UsedImplicitly]
         public static int CakeConvert(string[] args, [CanBeNull] AbsolutePath rootDirectory, [CanBeNull] AbsolutePath buildScript)
         {
+            PrintInfo();
+            Logging.Configure();
             Telemetry.ConvertCake();
             ProjectModelTasks.Initialize();
 
@@ -47,12 +49,12 @@ namespace Nuke.GlobalTool
                 }.JoinNewLine());
 
             Host.Debug();
-            if (!UserConfirms("Continue?"))
+            if (!PromptForConfirmation("Continue?"))
                 return 0;
             Host.Debug();
 
             if (buildScript == null &&
-                UserConfirms("Should a NUKE project be created for better results?"))
+                PromptForConfirmation("Should a NUKE project be created for better results?"))
             {
                 Setup(args, rootDirectory: null, buildScript: null);
             }
@@ -86,7 +88,7 @@ namespace Nuke.GlobalTool
             Host.Information("Found .cake files:");
             cakeFiles.ForEach(x => Host.Debug($"  - {x}"));
 
-            if (UserConfirms("Delete?"))
+            if (PromptForConfirmation("Delete?"))
                 cakeFiles.ForEach(FileSystemTasks.DeleteFile);
 
             return 0;