+++ website/src/pages/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import clsx from 'clsx';
+import Layout from '@theme/Layout';
+import Link from '@docusaurus/Link';
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+import styles from './index.module.css';
+
+export default function Home() {
+  const {siteConfig} = useDocusaurusContext();
+  return (
+    <div>
+      <h1>Something exciting is building...</h1>
+      <h3>Come back soon!</h3>
+    </div>
+  );
+}